Update 01_selection_sort.py
The incumbent version of the file is missing a statement that re-assigns the value of smallest in findSmallest when a lower value is detected.  See pg. 35 of the original text-I have added the statement.  Coincidentally, the incumbent program will produce the correct result because of the order of the items in the sample input data.  If the data are in a different order (switch the 10 and 3, for example), the output is not properly sorted unless the code I have indicated is added.  Thanks for your consideration!